Add non-shell commands to SSH port supplier
Extends the SSH port supplier to allow commands to be run without starting
a shell and allocating a PTY on the remote host.
